Colorjam hue warp presets

Colorjam hue warp presets

Usage

colorjam_presets(preset = NULL, ...)

Arguments

preset

NULL to return a character vector of all recognized preset names, or character string to return specific data associated with a recognized preset name.

...

additional arguments are ignored.

Value

character vector of recognized colorjam presets, or when preset is provided, it returns a list with elements:

h1: HCL color hue indicating the actual hue to be used when generating a color.

h2: HCL color hue indicating the virtual hue assigned to each actual hue in h1.

direction: with 1 indicating forward (increasing, clockwise) progression around the color wheel, and -1 indicating reverse (decreasing, counter-clockwise) progression around the color wheel.

default_step: character string with name matching a value returned by colorjam_steps()

optional attribute attr(x, "description") which may contain a text description for each color preset

Details

colorjam_presets(): list the names of available colorjam presets, or when a preset name is provided, it returns hue warp data as a list with names "h1" and "h2" suitable for use by h2hw() and hw2h().

See also