Get hue color warp options

Get hue color warp options used to convert to and from warped hues

h2hwOptions(
  h1 = getOption("h2hw.h1"),
  h2 = getOption("h2hw.h2"),
  preset = getOption("colorjam.preset", "custom"),
  preset_names = c("custom", "none", "dichromat", "rgb", "ryb", "ryb1", "ryb2", "ryb3"),
  default_preset = "dichromat",
  reset = FALSE,
  setOptions = c("ifnull", "TRUE", "FALSE"),
  h1default = NULL,
  h2default = NULL,
  verbose = FALSE,
  ...
)

Arguments

h1

numeric vector of color hue values which represent the "from" hue angles, also referred to as "hue warp" or "hw" values. One example of a typical operation: one may want to know the R hue for a particular red-yellow-blue hue. In this scenario the h1 "hue warp" is the red-yellow-blue hue, and the "hue" is the R typical hue used in colorspace::polarLUV() grDevices::hcl(), and farver::convert_colour(). For evenly-spaced red-yellow-blue colors, one would define a sequence of "hue warp" values from 0 to 360, then convert them to the default hue used by other R functions.

h2

numeric vector of color hue values which represent the "to" hue angles, also referred to as "hue" or "h".

preset

character string indicating whether to define h1, and h2 values based upon named presets:

"custom" uses values defined in options("h2hw.h1") and options("h2hw.h2") if they exist, otherwise default_preset.

"ryb" red-yellow-blue; "dichromat" color wheel intended to be color-blind friendly based upon simulated output as from the dichromat::dichromat() package;

"rgb" or "none" both use the hue defined in R as red-green-blue;

"ryb1","ryb2","ryb3" experimental red-yellow-blue alternatives that continue to modify the effect of the huge blue-green range of hue angles.

preset_names

character vector of valid preset values, included only to make the values visible in the function arguments.

default_preset

character string indicating which value in preset should be used as the default when reset=FALSE and h1 and/or h2 are not defined in options().

reset

logical whether to reset h1 and h2 values to the default values as defined in h1default and h2default. When reset=TRUE all other preset and default_preset arguments are ignored.

setOptions

character or logical indicating whether to update options() for "h2hw.h1" and "h2hw.h2". When "ifnull" the options() are only updated if they were previously NULL.

h1default, h2default

deprecated and ignored in this version of the function.

verbose

logical whether to print verbose output

Value

list with names h1 and h2 containing numeric vectors of hues between 0 and 360.

Details

This function retrieves, and/or defines, the h1 and h2 hue vectors used to convert between h1 "hue warp", and h2 "default hue". The "default" hue is used by standard R functions such as colorspace::polarLUV() grDevices::hcl(), and farver::convert_colour(), based upon the red-green-blue color wheel in typical computer video displays.

The h1 values may represent colors on red-yellow-blue color wheel hue, while h2 represents the corresponding hue in R "default hue" space. The conversion is non-linear, therefore it warps (bends) the color range accordingly.

The output values from this function are used with approx_degrees() which either performs the conversion, or returns a function to perform the conversion.

There are three options() used when available, intended to help custom states to have persistence across other function calls.

options("colorjam.preset") which keeps track of the current color warp preset.

options("h2hw.h1") defines the h1 value.

optinos("h2hw.h2") defines the h2 value.

Note that options("h2hw.h1") and options("h2hw.h2") are only used when preset="custom", otherwise the named preset will be used as first priority.

To disable the warped hue mechanic, set preset="none" or define h1 = h2 so the conversion is always 1:1.

When preset="custom" and h1 and h2 are not defined, the preset is set to the value of default_preset.

See also

Examples

h2hwOptions()
#> $h1
#> [1]   8  30  65 120 200 240 260 280 330
#> 
#> $h2
#> [1]   0.0  79.1 118.7 118.7 118.7 126.6 185.9 304.6 344.2
#> 
h2hw(60)
#> [1] 113.0429

h2hwOptions(h1=c(0, 60,120,240,300,360),
   h2=c(0,120,180,240,280,360))
#> $h1
#> [1]   8  30  65 120 200 240 260 280 330
#> 
#> $h2
#> [1]   0.0  79.1 118.7 118.7 118.7 126.6 185.9 304.6 344.2
#> 
h2hw(300)
#> [1] 320.44