Jam color palette for ggplot2

Jam color palette for ggplot2

jam_pal(
  type = "seq",
  palette = 1,
  direction = 1,
  invert = FALSE,
  darkFactor = 1,
  sFactor = 1,
  darken = FALSE,
  alpha = 1,
  useGrey = 20,
  preset = getOption("colorjam.preset", "dichromat2"),
  step = getOption("colorjam.step", "v24"),
  ...
)

Arguments

type

character string indicating the colors are sequential "seq".

palette

integer value indicating the categorical palette to use, intended to provide variety in the color assignment. (Not yet implemented.)

direction

integer indicating whether to reverse the color assignment, either 1 for the default forward assignment, or -1 for reverse assignment. Any negative value will reverse the colors.

invert

logical indicating whether to return corresponding contrasting colors, for example for text labels, typically either "white" or "black" as defined by jamba::setTextContrastColor().

darkFactor, sFactor

numeric passed to jamba::makeColorDarker() for optional adjustment of the color, by darkness and saturation, respectively. Intended when using scale_color_jam() and scale_fill_jam() where you want the color value to be lighter or darker than the fill color, a useful effect for outlines.

darken

logical indicating whether to apply default values for darkFactor and sFactor to darken the resulting colors.

alpha

numeric value indicating the alpha transparency, on a scale of 0 (transparent) to 1 (non-transparent).

useGrey

integer value between 0 and 100 indicating the grey value, as sent to jamba::setTextContrastColor(), used only when invert=TRUE.

...

additional arguments are passed to rainbowJam().

See also