Colorjam R-shiny App

Colorjam R-shiny App for color selection and custom adjustments

launchColorjamShiny(
  ...,
  envir = globalenv(),
  options = list(width = 1200, host = "0.0.0.0", port = 8080),
  verbose = TRUE
)

Arguments

...

additional arguments are passed to shiny::shinyApp() when the argument name is recognized by shiny::shinyApp().

envir

environment

options

list passed to shiny::shinyApp(). The most common options are:

host with either a specific IP address which then requires all incoming web requests to match that IP address, or host="0.0.0.0" to accept web requests for any host.

port with specific numeric web port, or port=NULL to assign a random port number when the R-shiny app is started.

width with a user-configurable width for the R-shiny app, useful when graphical components should expect at least a minimum width.

verbose

logical indicating whether to print verbose output.

Value

An object that represents the R-shiny app. Printing the object or passing it to shiny::runApp() will run the app.

See also

Examples

if (interactive()) {
   launchColorjamShiny();
}