Jam default theme for ggplot2

Jam default theme for ggplot2

theme_jam(theme_default = theme_bw, base_size = 18,
  grid.major.size = 0.5, grid.minor.size = 0.25,
  strip.background.colour = "grey30",
  strip.background.fill = "lightgoldenrod1",
  panel.grid.major.colour = "grey60",
  panel.grid.minor.colour = "grey80", axis.text.x.angle = 60,
  blankGrid = FALSE, blankXgrid = FALSE, blankYgrid = FALSE,
  resetTheme = TRUE, verbose = FALSE, ...)

Arguments

theme_default
function representing a ggplot2 theme.

base_size
default font point size, used for scaling the overall text sizes larger or smaller.

grid.major.size, grid.minor.size
the line width for the major and minor grid lines, respectively. Set to 0 to suppress either.

strip.background.colour, strip.background.fill
color for the border and strip background itself when ggplot2 is using a faceted layout.

panel.grid.major.colour, panel.grid.minor.colour
colors for the major and minor grid lines, respectively.

axis.text.x.angle
numeric degrees to rotate the x-axis labels, apparently starts at 0 (horizontal) and goes counter-clockwise (to the left.)

blankGrid, blankXgrid, blankYgrid
logical indicating whether to have a blank grid for everything, major, or minor axis lines, respectively. Intended to make it fast and easy to remove all gridlines.

resetTheme
logical whether to call the function theme_default which essentially resets (replaces) all previous settings with those defined in the theme function. If FALSE then only the specific settings defined in this function will be applied.

verbose
logical indicating whether to print verbose output.

...
additional arguments are passed to ggplot2::theme() in order to allow custom settings beyond what this function provides.

Details

This function applies some default theme settings to ggplot2, mainly taking away the default grey newspaper background color, also rotates the x-axis label text to 60 degrees, to accomodate longer labels without overlaps.