Colorjam chroma/luminance steps

Colorjam chroma/luminance steps to adjust a series of color hues to visibly distinct categorical colors.

colorjam_steps(step = NULL, ...)

Arguments

...

additional arguments are ignored.

steps

NULL to return a character vector of all recognized steps, or character string to return specific data associated with a recognized steps name.

Value

character vector of recognized colorjam step names, or when step is provided, it returns a list with elements "C" and "L".

Details

colorjam_steps(): list the names of available colorjam steps, or when a step name is provided, it returns a list with elements "C" and "L".

See also