Convert standard hue to warped hue

Convert standard hue to warped hue using the hue warp vectors

h2hw(h, h1 = h2hwOptions()$h1, h2 = h2hwOptions()$h2, ...)

Arguments

h
numeric vector of color hues between 0 and 360. These hues do not need to be in sequential order.

h1, h2
vector of color hues, which by default are defined in h2hwOptions(), but allowed here in cases where the global options should be overridden but not modified.

Value

numeric vector of hue values after applying the hue warp operation.

numeric vector of warped color hues.

Details

This function is intended to convert from a vector of hue values to the warped hues defined by the vectors returned by h2hwOptions(). The intent is to convert colors in RGB space into RYB space by default, which substantially improves several other color manipulations, such as selection of categorical colors, and color blending.

Note the input hue is considered the "standard" color hue as defined by the colorspace::polarLUV() function, ranging between 0 and 360. By this standard, 0 is defined as red, 120 is defined as green, and 240 is defined as blue.

The default mappings convert RGB (red, green, blue) to RYB (red, yellow, blue). The color yellow has hue=60 in RGB space, so the call to hw2h(60) results in 120, which is the hue in RYB space.

See also

Examples

## Yellow when using an RGB color wheel is 60 degrees,
## but on an RYB color wheel is 120 degrees.
h2hw(60);

#> [1] 120

# RGB colors are convenient, but are not ideal especially when blending
# colors. Note that blue and yellow have hues that differ by exactly 180
# degrees, meaning a hue average is as likely to be purple as green.
huesBY <- jamba::col2hcl(c("blue", "yellow"))["H",];
huesBY;

#> [1] 265.87278  85.87351

warpedHuesBY <- h2hw(huesBY);
warpedHuesBY;

#> [1] 265.8728 145.8735

warpedHues <- h2hw(hues);

#> Error in approx(x = h1, y = h2, xout = (h%%360)): object 'hues' not found