Get hue color warp options

Get hue color warp options used to convert to and from warped hues

h2hwOptions(
  h1,
  h2,
  h1default = c(0, 60, 120, 240, 360),
  h2default = c(0, 120, 180, 240, 360),
  preset = c("dichromat", "none", "ryb", "rgb", "ryb2"),
  reset = FALSE,
  verbose = FALSE,
  ...
)

Arguments

h1
NULL or numeric vector of color hue values, sequential between 0 and 360.

h2
NULL or numeric vector of color hue values, sequential between 0 and 360.

h1default, h2default
numeric vector of color hue values to use when h1 or h2 are NULL, respectively. The default values are included here as a function parameter here for visibility.

preset
character string indicating whether to define the h1, and h2 values based upon named presets: "none" no change; "ryb" red-yellow-blue; "dichromat" color wheel that omits green, therefore provides relatively even spacing of color-blind-friendly colors based upon simulated output as produced by dichromat::dichromat() package; "rgb" red-green-blue which is the default color wheel used in R; "ryb2" alternate red-yellow-blue that slightly over-emphasizes yellow at the expense/benefit of even less green.

reset
logical whether to reset h1 and h2 values to the defaults, as defined in h1defaults and h2defaults, respectively.

verbose
logical whether to print verbose output

Value

list with names h1 and h2 containing numeric vectors of hues between 0 and 360.

Details

This function simply retrieves the vector of hues in normal hue color space, and warped color space. It is intended to convert from something like a color wheel defined by RGB (red, green blue) where red=0, green=120, and blue=240, to a color wheel like RYB (red, yellow, blue) where red=0, yellow=120, and blue=240, (green becomes 180). This conversion is non-linear, meaning it warps (bends) the colors.

The purpose of using options() to store the color hue is to make it easy to re-use the settings across multiple function calls. The related reason for this function h2hwOptions() is to make it clear the default values, and make it clear how to update and review the current settings.

Together the vector h1 and h2 are used by stats::approx() to convert from the value ranges in h1 to the corresponding value ranges in h2. Using this mechanism, one could define ranges which effectively remove entire slices of the color wheel.

To disable the warped hue mechanism, set h2 equal to h1 for example h2hwOptions(h2=h2hwOptions()$h1). Doing so will cause stats::approx() to interpret every hue as a 1:1 relationship with the warped hue.

See also

Examples

h2hwOptions()

#> $h1
#>  [1]   0   8  30  65 120 200 240 260 280 330 360
#> 
#> $h2
#>  [1]   0.0   0.0  79.1 118.7 118.7 118.7 126.6 185.9 304.6 344.2 360.0
#> 

h2hw(60)

#> [1] 113.0429

h2hwOptions(h1=c(0, 60,120,240,300,360),
   h2=c(0,120,180,240,280,360))

#> $h1
#>  [1]   0   8  30  65 120 200 240 260 280 330 360
#> 
#> $h2
#>  [1]   0.0   0.0  79.1 118.7 118.7 118.7 126.6 185.9 304.6 344.2 360.0
#> 

h2hw(300)

#> [1] 320.44