Convert warped hue to standard hue

Convert warped hue to standard hue using the hue warp vectors

hw2h(
  h,
  h1 = NULL,
  h2 = NULL,
  preset = getOption("colorjam.preset", "custom"),
  ...
)

Arguments

h

numeric vector of color hues between 0 and 360. These hues do not need to be in sequential order.

h1, h2

numeric vector of color hues, which by default are defined in h2hwOptions(), but allowed here in cases where the global options should be overridden but not modified.

Value

numeric vector of color hues after applying the transformation from h2 to h1.

numeric vector of hue values after applying the hue warp operation.

Details

This function is intended to convert from a vector of warped hue values to the hues defined by the vectors returned by h2hwOptions()`. The intent is to convert colors in RYB space into RGB space by default.

One example of input would be to supply a uniformly spaced set of color hues, and convert them to "standard" hues by HCL standards. It has the effect of presenting categorical colors, using a non-linear set of "standard" hue values.

The default mappings convert RYB (red, yellow, blue) to RGB (red, green, blue). The color yellow has hue=120 in RYB space, so the call to h2hw(120) results in 60, which is the hue in RGB space.

See also

Examples

# It can be useful to create a uniform sequence of angles in warped
# hues, which are visually more uniform than those using an RGB color wheel,
# then convert those hues to standard color hues.
warpedHues <- seq(from=0, to=330, length.out=12);
warpedHues;
#>  [1]   0  30  60  90 120 150 180 210 240 270 300 330
hues <- hw2h(warpedHues);
hues;
#>  [1]   8.00000  16.34387  24.68774  39.63384 240.38690 249.31548 258.24405
#>  [8] 264.06066 269.11542 274.17018 279.22494 312.07071